Removed all text styles apart from body text style. Moved convert to text style to TextbitEditor for use in plugins
