+++ src/components/core/TextElement/style.css
@@ -5,33 +5,7 @@
   letter-spacing: -0.003em;
 }
 
-.core\/text .textbit-h1 {
-  font-size: 200%;
-  font-weight: bold;
-  font-family: var(--font-family-sans-serif, InterVariable, "Helvetica Neue", Helvetica, sans-serif);
-}
-
-.core\/text .textbit-h2 {
-  font-size: 125%;
-  font-weight: bold;
-  font-family: var(--font-family-sans-serif, InterVariable, "Helvetica Neue", Helvetica, sans-serif);
-}
-
-.core\/text .textbit-preamble {
-  font-weight: bold;
-}
-
-.core\/text .textbit-dateline {
-  display: inline-block;
-  padding: 3px 6px;
-  background-color: hsl(var(--secondary));
-  color: hsl(var(--secondary-foreground));
-  font-weight: bold;
-  font-size: 90%;
-  font-family: var(--font-family-sans-serif, InterVariable, "Helvetica Neue", Helvetica, sans-serif);
-}
-
-.core\/text .textbit-unknown {
+.core\/text .core\/text-unknown {
   font-style: italic;
   text-decoration: line-through;
   opacity: 0.7;
