+++ src/components/core/index.ts
@@ -1,8 +1,6 @@
 import { TextElement } from './TextElement'
 import { Bold, Italic, Underline } from "./Leafs"
 import { Navigation } from "./Navigation"
-import { OEmbed } from "./OEmbed"
-import { Image } from "./Image"
 import { Loader } from "./Loader"
 
 
@@ -10,9 +8,10 @@ export const StandardPlugins = [
   TextElement,
   Bold,
   Italic,
-  Underline,
+  Underline
+]
+
+export const basePlugins = [
   Navigation,
-  OEmbed,
-  Image,
   Loader
 ]
